⚡ Bolt: optimize sensor magnitude float arithmetic in StrokeDetector
Eliminate unnecessary Float -> Double -> Float type conversions in StrokeDetector's onSensorChanged callback by utilizing single-precision float sqrt math directly.

+++ app/src/main/java/com/openkayak/app/service/StrokeDetector.kt
@@ -81,7 +81,9 @@ class StrokeDetector(context: Context) : SensorEventListener {
         val ay = event.values[1]
         val az = event.values[2]
 
-        val currentAcc = sqrt((ay * ay + az * az).toDouble()).toFloat()
+        // Performance Optimization: Use Float overload for sqrt directly to avoid
+        // converting Float -> Double -> Float on every high-frequency sensor event (20-50Hz).
+        val currentAcc = sqrt(ay * ay + az * az)
 
         val filteredAcc = previousAcceleration + 0.3f * (currentAcc - previousAcceleration)
